rewrite character counter code
The old character counter code had hoisted variables,
hardcoded values all over the place, and was generally
not well formatted. This update creates more modular code
and gives hook points for updating some of the values
it uses. I think this is step one, and we need to also
remove it's self initialization so options can get passed
across, or global defaults for the counter can be updated
to the developers liking.