fixed dropdowns misbehaving because of relocating dom
The issue is dropdowns are relocated in the dom to be right next to the element that is triggering the dropdown. If the surrounding dom has CSS elements that negatively affect the dropdown (like position: relative, or overflow: hidden), then the dropdown appears strange.

This fix gets the offset of the origin in relation to the entire page, and then it absolutely positions the dropdown from that.