Dynamically added list items are expandables too without reinitializing
This attaches one single event handler to an expandable list, but sets
the scope to the collapsible header. This way you attach one event
handler instead of potentially hundreds and it'll make sure any new
dynamically added list item is able to be collapsed without
reinitializing the component.

Here's a quick example of what I changed: https://jsfiddle.net/nyy085de/

Note: this only changes how the expandables work, not collapsibles. I
don't have the time to have a look at the rest, but I have the feeling
that this kind of thing happens on other parts of the projects; applying
event handlers to ONLY a cached element/list of things.