Modal enhancement to reset scroll position on exit
When a modal has dynamic content that can change with every load of the
modal (my use case is with different forms in angularjs) the scroll
position persists between loads. Added an option that is true by default
that will reset the modal content scroll position x,y to 0,0 when the
modal window is closed.