Make bower.json package name lower-case
Please consider following the convention of using lower-case for package names in bower, npm, etc. I lost a significant amount of time today debugging a strange problem that was ultimately caused by this package using a capitalized name but being referenced in the code by a lower-case one on a file system that is case-sensitive.