BUG: Labels animate floating after page render
When inputs are pre-filled, the label starts out overlapping
the text, and animates to the floating position after the DOM
is loaded. This fixes that for inputs and selects, but not
for textareas (those still need some initialization).

Also, a context can now be added to Materialize.updateTextFields,
(e.g. `Materialize.updateTextFields("#my-container")`) so that
the entire page does not need to be reinitialized. A new public
function is exposed as well - Materialize.updateTextField - to
allow allow users to update a single input element (which could
could be useful for frameworks like React, Angular, Aurelia, etc)

Removed unnecessary `$(document).ready` function when setting up
public Materialize functions. This allows users to start using
those functions without calling `$(document).ready` themselves.