Changes modal _durations for consistency.
Modals use in_duration and out_duration, but dropdowns and materialbox use inDuration and outDuration. JS convention is to use camelCase so this PR updates modals to follow convention of the language and rest of the framework.

Note, this is a breaking change. Anyone using modals and in_duration or out_duration will find their durations stop working. If you want, I can add backwards compatibility too, but I note that the project is not yet version 1, so breaking changes are allowed in semver. Let me know if that is not the case though.