jump to the tabs and it was necessary to scroll down to see the tab content.
The reason for that was that the content for the previously active tab was being hidden before
the content for the next tab was being shown, which was shortening the document.

As a fix the next tab content is shown first before the old tab content is hidden.

If the tabs are much different lengths then the document will still be shortened and there will
still be some movement, but only if the bottom of the shorter tab is above the bottom of
the screen