Introduce reponsive horizontal cards
By using the markup of the horizontal card with the class "responsive-horizontal" the card will display as a horizontal card only on medium an up devices. On small devices the card will be displayed as a regular card since horziontal cards do not make much sense on small devices anyway.