Allow carousel reinitialization
Previously it was not possible to reinitialize a carousel if items have changed by adding/removing .carousel-item nodes.

This patch allows reinitialization by calling .carousel() again. All events and classes are first removed and then the normal setup process continues.

I have also included in this patch that if a .carousel-item has the .active class, the carousel jumps to this item after (re-)initialization. The .active class can either be manually set or is set by the carousel itself when the user is scrolling to it.

This patch was necessary, because I pull updated items from a server and dynamically add/remove carousel items using Knockout.js. Now I only need to call $('#myCarousel').carousel() again.

I have tested this with the default carousel options and with { fullWidth: true }.