Carousel mobile tapping/vertical scrolling
The tap swipe up or swipe down events within the carousel are not propagating. It is also not possible to tap indicators or any controls/links contained inside the carousel items.

Apply this patch to restore the vertical scroll behaviour on mobile devices but leave it as is on desktop devices.