Change hiddenDiv.height() to .outerHeight()
The autosize feature for textareas does not work that well if one changes the style of the textarea to make use of padding and border-box.
I would suggest to simply change .height to .outerHeight for the hiddenDiv. It does not change anything to the current behavior, but one now can simply add the style also to the hiddenDiv, which then gets applied to the autosize automatically.

Another option would be to consider the textareas other css (beside the existing checks) and check for padding as well as box-sizing, and more. Also possible, but needs way more testing. The first approach would be a easier and faster fix.

Chose your favorite solution. One of them would be awesome!!