Add option to select remaining entry with enter
A small change that makes it possible to select and choose a last remaining entry by pressing enter. This was a requirement in my project and saves the user an unnecessary keystroke.
May not seem like much, but in an application that is used hundred of times a day it can make a real difference.