Allows element with .waves-effect to be centered
This fixes elements with `.waves-effect` class added to this. Currently they cannot be centered without overriding the `display: inline-block` style that was set here. I do not see a reason for setting that property as it appears to do nothing. Removing it allows for easier centerting of buttons with `.waves-effect` class added to them