Initial stab at future Ajax support
Add a callback hook just before data is consulted, leveraging
the logic that is already there (irrelevant keys, minlength, etc).

This needs more work to deal with requests typically longer
the time it takes for a user to type multiple keys, but the API
for the callback is very unlikely to change as all that logic can
be encapsulated in the callback.