Add use of format function on init if defined
When you define a custom format function for the datepicker, its not used on the init. The datepicker create the date object from the input value when no default date is defined, and uses the default format function from javascript that only accepts dates on the format "YYYY-MM-DD".
With these proposed change, on the init the datepicker will check if there is a custom format function defined and will use it insted of the default javascript one to load the date from the input value when no default date is defined.