Issue #5174 Make timepicker output format configurable.
Allow timepickers to be configured with a "format" option similar to
that used by datepickers. E.g. { format: "HH:mm" }
or { format: "hh:mm A" }. Accommodate 12-hour and 24-hour modes, and
allow users to specify output formats using any of the following
elements:
 - 24-hour format of the hour, with or without leading zeros
 - 12-hour format of the hour, with or without leading zeros
 - Minutes, with or without leading zeros
 - Seconds, with or without leading zeros (currently defaults to 0/00)
 - Uppercase or lowercase ante/post meridian (AM/PM)

This will allow the use of timepickers with HTML inputs of type "time"
in browsers that perform validation and reject the default format for
12-hour mode ("hh:mm A" - e.g. "12:45 AM").