Update input width in _chips.scss sass component
This is a minor commit which updates the width of input field from fixed 120px to fit-content with same !important weight. Also, the optional min-width and max-width are set to limit its size between 100px and 200px. With this change, the input placeholder (if limited to under 200px) will not get cropped, and it will also have a better fit into the input field.