feat(datepicker): add format method
To leverace browser own internationalization methods, it is useful
to add a format method instead of simple format string. That makes
things like this possible:

```js
M.Datepicker.init(elems, {
  format: date => new Intl.DateTimeFormat(locale).format(date)
});
```

Which would format the input according to the i18n rules defined
by the browser using the locale of the user.